Removed old constructor using double and having unused fields

+++ src/pipe/gui/graphicElements/PetriNetObjectWithLabel.java
@@ -42,13 +42,14 @@ public abstract class PetriNetObjectWithLabel extends PetriNetObject {
     protected void updateLabelLocation(boolean alignToGrid) {
         if(alignToGrid) {
             this.getNameLabel().setPosition(
-                    Grid.getModifiedX((int) (positionX + Zoomer.getZoomedValue(nameOffsetX, getZoom()))),
-                    Grid.getModifiedY((int) (positionY + Zoomer.getZoomedValue(nameOffsetY, getZoom())))
+                    Grid.getModifiedX(positionX + Zoomer.getZoomedValue(nameOffsetX, getZoom())),
+                    Grid.getModifiedY(positionY + Zoomer.getZoomedValue(nameOffsetY, getZoom()))
             );
         } else {
             this.getNameLabel().setPosition(
-                    ((int)(positionX + Zoomer.getZoomedValue(nameOffsetX, getZoom()))),
-                    ((int)(positionY + Zoomer.getZoomedValue(nameOffsetY, getZoom()))));
+                positionX + Zoomer.getZoomedValue(nameOffsetX, getZoom()),
+                positionY + Zoomer.getZoomedValue(nameOffsetY, getZoom())
+            );
         }
     }
